+++ NEWS.md
@@ -10,6 +10,10 @@
 
 * Group objects can be opened while supplying a Config object when 2.15.1 or newer is used (#535, #536)
 
+* For character column buffer allocations, the R function now accepts a `nullable` option (#537)
+
+* For standard buffer allocations, the R function now accepts `nullable` and `varnum` options (#538)
+
 ## Build and Test Systems
 
 * Testing for Groups reflect the stricter behavior in config setting requiring a close array (#530)